refactor: перенести фильтрацию фильмов в папку utils
